requirements.txt
gradio==4.44.0
transformers==4.43.3
torch
sentencepiece
accelerate
pypdf
trafilatura
gTTS
pydub
Pillow
langdetect
regex
numpy

